moved things around (maybe then I can talk to the task tab from the failures panel?)
