+++ ui/src/app/job-details/job-details.component.spec.ts
@@ -30,8 +30,8 @@ import {ActivatedRoute, Router} from '@angular/router';
 import {URLSearchParamsUtils} from '../shared/utils/url-search-params.utils';
 import {JobStatus} from "../shared/model/JobStatus";
 import {JobMetadataResponse} from '../shared/model/JobMetadataResponse';
-import {JobFailuresComponent} from "./failures/failures.component";
-import {JobFailuresTabComponent} from "./failures/failures-tab/failures-tab.component";
+import {JobFailuresComponent} from "./tasks/failures/failures.component";
+import {JobFailuresTabComponent} from "./tasks/failures/failures-tab/failures-tab.component";
 
 describe('JobDetailsComponent', () => {
