+++ ui/src/app/job-details/job-details.module.ts
@@ -22,8 +22,8 @@ import {JobResourcesTableComponent} from './resources/resources-table/resources-
 import {SharedModule} from '../shared/shared.module';
 import {TaskDetailsComponent} from './tasks/tasks.component';
 import {GcsService} from '../core/gcs.service';
-import {JobFailuresTabComponent} from './failures/failures-tab/failures-tab.component';
-import {JobFailuresComponent} from './failures/failures.component';
+import {JobFailuresTabComponent} from './tasks/failures/failures-tab/failures-tab.component';
+import {JobFailuresComponent} from './tasks/failures/failures.component';
 
 
 @NgModule({
