+++ ui/src/app/job-details/tasks/failures/failures-tab/failures-tab.component.css
