+++ ui/src/app/job-details/tasks/failures/failures-tab/failures-tab.component.ts
@@ -1,8 +1,8 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {JobMetadataResponse} from "../../../shared/model/JobMetadataResponse";
-import {FailureMessage} from "../../../shared/model/FailureMessage";
-import {ResourceUtils} from "../../../shared/utils/resource-utils";
-import {TaskMetadata} from "../../../shared/model/TaskMetadata";
+import {JobMetadataResponse} from "../../../../shared/model/JobMetadataResponse";
+import {FailureMessage} from "../../../../shared/model/FailureMessage";
+import {ResourceUtils} from "../../../../shared/utils/resource-utils";
+import {TaskMetadata} from "../../../../shared/model/TaskMetadata";
 
 @Component({
   selector: 'jm-failures-tab',
