+++ ui/src/app/job-details/tasks/failures/failures.component.css
