+++ ui/src/app/job-details/tasks/failures/failures.component.ts
@@ -1,7 +1,7 @@
-import {Component, Input, OnInit} from '@angular/core';
-import {FailureMessage} from "../../shared/model/FailureMessage";
-import {ResourceUtils} from "../../shared/utils/resource-utils";
-import {TaskMetadata} from "../../shared/model/TaskMetadata";
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {FailureMessage} from "../../../shared/model/FailureMessage";
+import {ResourceUtils} from "../../../shared/utils/resource-utils";
+import {TaskMetadata} from "../../../shared/model/TaskMetadata";
 
 @Component({
   selector: 'jm-failures',
