+++ ui/src/app/job-details/tasks/tasks.component.html
@@ -1,6 +1,6 @@
 <mat-expansion-panel class="content" expanded="true" [disabled]="true">
   <mat-expansion-panel-header class="title">Tasks</mat-expansion-panel-header>
-  <mat-tab-group [selectedIndex]="selectedTab">
+  <mat-tab-group [(selectedIndex)]="selectedTab">
     <mat-tab label="List View">
       <mat-table #table [dataSource]="dataSource">
         <!-- Task name column -->
