+++ ui/src/app/job-details/tasks/tasks.component.spec.ts
@@ -13,8 +13,8 @@ import {
 import {ClrIconModule, ClrTooltipModule} from '@clr/angular';
 
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
-import {JobFailuresComponent} from "../failures/failures.component";
-import {JobFailuresTabComponent} from "../failures/failures-tab/failures-tab.component";
+import {JobFailuresComponent} from "./failures/failures.component";
+import {JobFailuresTabComponent} from "./failures/failures-tab/failures-tab.component";
 import {JobMetadataResponse} from '../../shared/model/JobMetadataResponse';
 import {JobStatus} from '../../shared/model/JobStatus';
 import {SharedModule} from '../../shared/shared.module';
